refactor: separate types into type files
