+++ src/app/mypage/@roadmaps/in-progress/InProgressRoadmaps.tsx
@@ -12,9 +12,10 @@ import NotFound from '@/components/NotFound';
 import { ArticlesCardsGrid } from '@/components/shared/grid/ArticlesCardsGrid';
 import { SkeletonCardsGrid } from '@/components/shared/grid/SkeletonGrid';
 
-import { RoadMapInfo } from '@/app/roadmap/post/[...id]/@roadmapInfo/page';
 import { API_ROUTES, FAIL } from '@/constants';
 import { getApiResponse } from '@/utils/get-api-response';
+
+import { RoadMapInfo } from '@/types/post';
 
 const InProgressRoadmapList = () => {
   const { data: session, status } = useSession();
