+++ src/components/Image404.tsx
@@ -1,6 +1,6 @@
 import { ComponentPropsWithoutRef } from 'react';
 
-export function Illustration(props: ComponentPropsWithoutRef<'svg'>) {
+export function Image404(props: ComponentPropsWithoutRef<'svg'>) {
   return (
     <svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 362 145' {...props}>
       <path
