+++ src/types/user.ts
@@ -11,6 +11,32 @@ export interface UserData {
   nickname: string;
 }
 
+export interface Member {
+  id: number;
+  email: string;
+  nickname: string;
+  bio?: null | string;
+  avatarUrl?: null | string;
+  githubUrl?: null | string;
+  blogUrl?: null | string;
+  baekjoonId?: null | string;
+  provider?: string;
+  accessToken?: string;
+}
+
+// interface Member {
+//   id: number;
+//   email: string;
+//   nickname: string;
+// bio?: null | string;
+// avatarUrl?: null | string;
+// githubUrl?: null | string;
+// blogUrl?: null | string;
+// baekjoonId?: null | string;
+// provider?: string;
+// accessToken?: string;
+// }
+
 export interface UserDataResponse {
   userData: UserData | null;
 }
@@ -18,4 +44,14 @@ export interface UserDataResponse {
 export interface MyPageProps {
   info: ReactNode;
   roadmaps: ReactNode;
+}
+
+export interface AboutInfo {
+  [key: string]: unknown;
+  isJoined: boolean;
+  isLiked: boolean;
+  joinCount: number;
+  likeCount: number;
+  member: Member;
+  updatedAt: string;
 }
